Many many cherry picks (#49)
* Monkey Martial Arts (#55583)

Moves martial arts to mob/living and lets monkeys use them.

Monkeys have got arms so they should be able to do cool martial arts.

* Add deathrattle implant for nuclear operatives (#55903)

Deathrattle implant, comes in box of 8, implant yourself and others, and
you'll get a little notification when they die.

Intended for nuke ops, price around the same as the codebook, since it's
a utility communication item, that actually "plans to fail", so it can't
be too expensive, because it doesn't help them win, just lets them know
how much they are losing by.

* Evil crab resprite! (#56215)

This PR resprites the evil crab in preparation for my evil crab rework.
The evil crab now has walking animation and a new fresh look.

* Refactor card scaling into an element for general item scaling (#56102)

The current TCG code had some code for scaling its cards down when they're on the ground and then scaling them back in hand/inventory. This element aims to preserve this functionality and to allow it work for other items.

While the TCG makes the cards smaller on the ground, this element allows for items to be scaled up OR down when on the floor or in inventory.

While this particular element has to do with scaling, I am looking at ways to expand this sort of icon change functionality to icon_state as well, but there are additional issues with blood decals needing to be redrawn and possibly vis_contents.

* Fixes Codex Cicatrix having infinite charges - Maintainer edition (#56307)

Gets rid of pointless text2num.
EK is a path and not an object. Fixes it being used as if it was an object through using initial.

* Admins transforming a humanoid into an AI can now choose to not move them to the satellite.  (#56075)

Adds a Yes/No/Cancel button to the AI transform proc that admins use on whether the new AI should be moved to the satellite or not.

Sometimes admins want to make an AI somewhere off-station for whatever reason. Also makes testing things with AIs less of a hassle.

* Attempts to fix addictions being utterly broken and almost impossible to obtain. (#56302)

* Patches gene shears click stack exploit (#56301)

Fixes #55390 by checking the plant's heath again after the input is done and before the gene is removed. I'm not sure if it's the cleanest way to get around stacking inputs to exploit the health, but it works?

* Fixes for the invasive spreading trait (#56300)

    Fixes invasive spreading overtaking same type plants (It was checking for dead plants instead of alive ones)
    Fixes invasive spreading going through glass (It only checked plants one step away, regardless of adjacency)
    Better var names (HY -> spread_tray, H -> our_tray)
    Slight visual feedback when plants are destroyed
    TRUE / FALSE instead of 1 / 0

* Adds Noogies (Perfect for bullying Moths) (#56286)

* i HATE mothblocks

Co-authored-by: Jack LeCroy <3073035+jacklecroy@users.noreply.github.com>
Co-authored-by: coiax <yellowbounder@gmail.com>
Co-authored-by: Krysonism <49783092+Krysonism@users.noreply.github.com>
Co-authored-by: Koshenko <53068134+Koshenko@users.noreply.github.com>
Co-authored-by: Timberpoes <silent_insomnia_pp@hotmail.co.uk>
Co-authored-by: zxaber <37497534+zxaber@users.noreply.github.com>
Co-authored-by: MrMelbert <51863163+MrMelbert@users.noreply.github.com>
Co-authored-by: Ryll Ryll <3589655+Ryll-Ryll@users.noreply.github.com>