Splits node design IDs into multiple lines in alphabetical order with trailing commas (tgstation/tgstation#59470)
This helps to prevent merge conflicts, as well as it making it easier to read. This has proven to be important with circuits, where multiple people want to add new components, but will end up inevitably conflicting on each other.

Also, aligns with the contribution guideline to not use tab alignment.