Fixes mouse hover triggering a do_after with a non-instantly equippable item in inv slot. (tgstation/tgstation#59548)
Mousing over an inventory slot with an item in-hand can trigger a call to equip_delay_self_check.

If the item has an equip_delay_self and bypass_equip_delay_self is FALSE, then it acts like it's going to equip the item and starts a do_after.

I think the only two items that do this are /obj/item/clothing/gloves/cargo_gauntlet and /obj/item/clothing/suit/straight_jacket

I make sure the call to can_equip back in the /atom/movable/screen/inventory/add_overlays section of the call stack calls with bypass_equip_delay_self = TRUE to prevent this.