Stimformation requirements list fix + some small heads-up on the requirements list (tgstation/tgstation#59599)
stimformation has a min() proc hooked to the gas formed that accounts for the amount of plasma present inside it, while not explicitly stating that in the requirements list. This PR would make stimformation/react() not get called by gas_mixture/react() when no gas will be formed.
Also added a short doc-comment to the requirements list in case someone else wanted to add new requirement interactions.