Replaces most HTML spans with macros
This is a manual compilation of the proc PR from TG at
tgstation/tgstation#59645
and the direct application of tgstation/tgstation#59685

Also refactors the spans to be UPPERCASE, to not get mixed up with
functions/procs. Helps a bit with visual cue load on the editors.

Additionally, renames some speech spans that were macro'd as SPAN_X to
SPEECH_SPAN_X.
That system should get overhauled at some point...